requirements.txt
datasets==2.20.0
faiss-cpu==1.8.0
openai==1.23.0
numpy==1.26.4
pyarrow==15.0.2
tiktoken==0.6.0
requests>=2.32.2
tqdm==4.66.4
pandas==2.2.2
httpx<0.27 
